Create re-usable entries table
- Use Entries table on Dashboard and Country page
- Use MyResources on Country page
