+++ src/components/EntriesTable/styles.css
@@ -1,0 +1,7 @@
+.table {
+    width: 100%;
+
+    * {
+        border: 0!important;
+    }
+}
