+++ src/components/MyResources/ResourceItem/index.tsx
