+++ src/components/MyResources/ResourceItem/styles.css
