+++ src/components/MyResources/myResources.interface.ts
